Move out ApTypeLabels
It feels wrong having to reach into an Apply page to get these labels
given they're a cross-application concern that is used in a lot of
places. Lets move them out to the top level of the ./server/utils folder
