+++ server/utils/apTypeLabels.ts
@@ -1,0 +1,10 @@
+import { ApType } from '../@types/shared'
+
+export const apTypeLabels: Record<ApType, string> = {
+  normal: 'Standard AP',
+  pipe: 'Psychologically Informed Planned Environment (PIPE)',
+  esap: 'Enhanced Security AP (ESAP)',
+  rfap: 'Recovery Focused AP (RFAP)',
+  mhapElliottHouse: 'Specialist Mental Health AP (Elliott House - Midlands)',
+  mhapStJosephs: 'Specialist Mental Health AP (St Josephs - Greater Manchester)',
+}
